Added smaller LUTs to repo
This is an example of files output using the generateLUT function in /code/utilities. Other larger LUTs are needed to reproduce the figures in the paper, but they are too big to fit here. If you run the "runAllData" script it will generate the larger LUTs but will take a while (8-ish hours). If you have a large file transfer service I can send you the pre-generated LUTs. Contact roblyer@bu.edu, or add an issue to this repo.